openoceancamera/requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
attrs==20.3.0
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.5
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cryptography==2.6.1
entrypoints==0.3
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.6
importlib-metadata==3.3.0
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.8
picamera==1.13
pluggy==0.13.1
py==1.10.0
pycrypto==2.6.1
PyGObject==3.30.4
pylint==2.3.1
pyparsing==2.4.7
pytest==6.2.1
python-apt==1.8.4.1
pyxdg==0.25
requests==2.21.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
smbus==1.1.post2
ssh-import-id==5.7
toml==0.10.2
tsl2561==3.4.0
typed-ast==1.4.0
typing-extensions==3.7.4.3
urllib3==1.24.1
Werkzeug==0.15.5
wrapt==1.11.2
zipp==3.4.0
